refactor isValidDropPosition to be generalizable to all dropped elements
also refactor the func to define valid drop positions instead of invalid
